+++ lib/streaker.rb
@@ -1,6 +1,7 @@
-require 'streak-ruby'
+require 'http'
 
 require 'streaker/version'
 require 'streaker/configuration'
+require 'streaker/api'
 require 'streaker/box'
 require 'streaker/box_not_found_error'