+++ lib/streaker/configuration.rb
@@ -2,22 +2,16 @@
 module Streaker
   # The configuration singleton.
   class Configuration
-    attr_reader :api_key
     attr_accessor :pipeline_keys,
                   :stage_keys,
-                  :field_keys
+                  :field_keys,
+                  :api_key
 
     def initialize
       @api_key = ''
       @pipeline_keys = {}
       @stage_keys = {}
       @field_keys = {}
-    end
-
-    def api_key=(value)
-      @api_key = value
-
-      Streak.api_key = value
     end
   end