+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2018-08-07
+
+Features:
+- Replace `streak-ruby` dependency by `http`.
+- Make only one call to update a box instead of one per attribute.
+- Create a box with persons assigned.
+
 ## [1.0.0] - 2017-12-11
 
 First release! :tada: